# Make Box2d::contains and Box2d intersects faster