@nical 2024-05-19
The speedup for `contains_box`:

Before:
```
box2d contains_box f32  time:   [436.69 ms 436.97 ms 437.26 ms]
```

After:
```
box2d contains_box f32  time:   [244.31 ms 244.51 ms 244.72 ms]                                   
change: [-44.102% -44.044% -43.981%] (p = 0.00 < 0.05)
```